refactor: 경험 추가 검증 로직 분리
