+++ src/app/(pages)/experience/add/_components/ExperienceAddPageContent.tsx
@@ -11,15 +11,10 @@ import type {
 import { ExperienceAddProgress } from '@/app/(pages)/experience/add/_components/ExperienceAddProgress';
 import { ExperienceAddStepContent } from '@/app/(pages)/experience/add/_components/ExperienceAddStepContent';
 import { EXPERIENCE_ADD_STEPS } from '@/app/(pages)/experience/add/_constants/experienceAddSteps';
-import { CORE_EXPERIENCE_FIELDS } from '@/app/(pages)/experience/add/_constants/experienceCoreQuestions';
-import { EXPERIENCE_TYPE_FIELD_GROUPS } from '@/app/(pages)/experience/add/_constants/experienceTypeOptions';
 import {
   createEmptyBasicInfoForm,
   createEmptyCoreInfoForm,
   createEmptyResultInfoForm,
-  type ExperienceAddBasicInfoForm,
-  type ExperienceAddCoreInfoForm,
-  type ExperienceAddResultInfoForm,
 } from '@/app/(pages)/experience/add/_types/experienceAddForm';
 import {
   mapAnalyzeResponseToBasicInfoForm,
@@ -31,6 +26,11 @@ import {
   clearExperienceAddPdfDraft,
   getExperienceAddPdfDraft,
 } from '@/app/(pages)/experience/add/_utils/experienceAddPdfDraftStorage';
+import {
+  isBasicInfoComplete,
+  isCoreInfoComplete,
+  isResultStepComplete,
+} from '@/app/(pages)/experience/add/_utils/experienceAddValidation';
 import { ErrorDialog } from '@/components/common/ErrorDialog';
 import { ChevronLeftIcon } from '@/components/common/icons/ChevronLeftIcon';
 import { Button } from '@/components/ui/button';
@@ -236,12 +236,7 @@ export function ExperienceAddPageContent() {
               이전
             </Button>
           )}
-          <Button
-            type="button"
-            className="w-40"
-            disabled={isNextStepDisabled}
-            onClick={goNextStep}
-          >
+          <Button type="button" className="w-40" disabled={isNextStepDisabled} onClick={goNextStep}>
             {currentStepIndex === EXPERIENCE_ADD_STEPS.length - 1 ? '저장하기' : '다음'}
           </Button>
         </footer>
@@ -257,41 +252,4 @@ export function ExperienceAddPageContent() {
       />
     </div>
   );
-}
-
-function isBasicInfoComplete(basicInfo: ExperienceAddBasicInfoForm) {
-  if (!basicInfo.type) return false;
-
-  return EXPERIENCE_TYPE_FIELD_GROUPS[basicInfo.type].every((fieldGroup) =>
-    fieldGroup.fields.every((field) => hasText(basicInfo[field.name])),
-  );
-}
-
-function isCoreInfoComplete(coreInfo: ExperienceAddCoreInfoForm) {
-  return CORE_EXPERIENCE_FIELDS.every((field) => hasText(coreInfo[field.name]));
-}
-
-function isResultStepComplete({
-  basicInfo,
-  coreInfo,
-  resultInfo,
-}: {
-  basicInfo: ExperienceAddBasicInfoForm;
-  coreInfo: ExperienceAddCoreInfoForm;
-  resultInfo: ExperienceAddResultInfoForm;
-}) {
-  return (
-    isBasicInfoComplete(basicInfo) &&
-    isCoreInfoComplete(coreInfo) &&
-    hasTag(resultInfo.skillTags) &&
-    hasTag(resultInfo.competencyTags)
-  );
-}
-
-function hasText(value: string | null | undefined) {
-  return (value ?? '').trim().length > 0;
-}
-
-function hasTag(tags: string[]) {
-  return tags.some(hasText);
 }
